eventhub-client/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio

